Fix time.sleep without movement
I had the problem that there was a time.sleep in the output script, even though the mouse was not moved. 
I moved the check, if there is enough mouse movement to the on_move function. Like this it should be cleaner and there is only a sleep if the mouse is really moving :)